Experiment with interfaces in `doIdleTask`
This is an experiment to attempt to make the primary function in
`RifLoaderIdleTasks` extreamly readable.

The goal is that anyone can read the content of `doIdleTask` and
understand at a high level what they should expect when using that
function.

To that end an interface has been created for a Task, allowing the
reader to see that there are tasks without having to care initially what
they are, while still allowing them to find out if they continue reading.

There may be other benefits and drawbacks to this approach.